+++ CHANGELOG.md
@@ -7,6 +7,24 @@ Features:
 - Add `link` props to `Field.Label`
 - Add `Field.Password` component
 
+## [1.23.0] - 2019-03-20
+
+Features:
+- Add `Field.Checkbox`.
+- Export `FLUID`, `DEFAULT`, `TINY` and `BIG` css from `Button`.
+
+Fix:
+- Update selected option style to `Select` component.
+
+## [1.22.0] - 2019-03-18
+
+Features:
+- Add `customSize` props to `HorizontalStroke`.
+- Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
+- Add `IconList`.
+- Add empty `Prev` and `Next` container to `NavTabCarousel` when
+  `NavTabCarousel.Prev` or `NavTabCarousel.Next` are empty.
+
 ## [1.21.0] - 2019-03-11
 
 Features: